border Tower
The tower can now calculate the range to the border - and the damage he
can do to the target. The idea is to deal with border pocker. Let him
move in till the and start shooting only when the target can not reach
the save exit - lets see how it works